requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
Django==2.0.13
django-countries==5.3.3
django-filter==1.1.0
django-smart-selects==1.5.4
django-smtp-ssl==1.0
djangorestframework==3.9.4
djangorestframework-filters==0.11.1
numpy==1.16.4
reportlab==3.5.23
tensorflow==1.13.1
tensorflow-estimator==1.13.0
xhtml2pdf==0.2.3
